+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 All notable changes to `subster/php-sdk` will be documented in this file.
+
+## Unreleased
+
+- Added optional checkout item quantity support via raw item arrays and `CreateCheckoutSessionItemData`.